requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
PySocks==1.7.1
python-telegram-bot==13.7
pytz==2021.1
six==1.16.0
tornado==6.1
tzdata==2021.1
tzlocal==3.0
emoji
clarifai-grpc==7.11.0
